backend/requirements.txt
Flask>=3.0.0
Flask-PyMongo>=2.3.0
PyJWT>=2.8.0
python-dotenv>=1.0.1
requests>=2.31.0
